feat: Add AutowiredEvent decorator for alarm-created event
This commit adds a new decorator called AutowiredEvent, which is used to automatically register event classes in the EventClsRegistry. The decorator is applied to the AlarmCreatedEvent class in the alarms domain, allowing it to be registered and used for deserialization in the EventDeserializer class.

The AutowiredEvent decorator improves the scalability and maintainability of the event handling system by providing a centralized way to register event classes.